'ajustes'

+++ web/todosAtendimentos.jsp
@@ -40,47 +40,44 @@
             <div class="container">
                 <a class="btn btn-outline-dark" href="novoatendimento.jsp">Adicionar um novo atendimento</a>
             </div>
-            <div id="loading">Carregando...</div>
-            <div class="container fluid" id="table">
-                <table class="table table-borderless table-hover mt-3">
-                    <thead class="thead-dark text-md-center">
-                        <tr>
-                            <th scope="col">Nome do Paciente</th>
-                            <th scope="col">Código</th>
-                            <th scope="col">Data de criação</th>
-                            <th scope="col">Situação</th>
-                            <th scope="col">Funcionário</th>
-                            <th scope="col">Ações</th>
-                        </tr>
-                    </thead>
-                    <tbody class="text-md-center">
-                        <%
-                            Atendimento a = new Atendimento();
-                            List<Atendimento> list = a.listAll();
-                            
-                            
-                            for(int i = 0; i < list.size(); i++){
-                                out.println("<tr>");
-                                out.println("<td>"+list.get(i).getNomepaciente()+"</td>");
-                                out.println("<td>"+list.get(i).getCodigo()+"</td>");
-                                out.println("<td>"+list.get(i).getData()+"</td>");   
-                                out.println("<td>"+list.get(i).getUsuario().getNome()+"</td>");   
-                                out.println("<td>"+list.get(i).getSituacao()+"</td>");
-                                out.println("<td>"); 
-                                out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=excluir&id="+list.get(i).getId()+"'>");
-                                out.println("<img src='img/excluir.png' title='excluir atendimento' width='20px' height='20px'>");
-                                out.println("</a>");
-                                out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=editar&id="+list.get(i).getId()+"'>");
-                                out.println("<img src='img/editar.png' title='editar atendimento' width='20px' height='20px'>");
-                                out.println("</a>");
-                                out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=detalhar&id="+list.get(i).getId()+"'>");
-                                out.println("<img src='img/informacao.png' title='detalhar atendimento' width='20px' height='20px'>");
-                                out.println("</a>");
-                                out.println("</td>");
-                                out.println("</tr>");
-                            }
-                            
-                        %>
+            <%
+                List<Atendimento> list = Atendimento.listAll();
+                if(list.size() > 0){%>
+                    <div id="loading">Carregando...</div>
+                    <div class="container fluid" id="table">
+                        <table class="table table-borderless table-hover mt-3">
+                            <thead class="thead-dark text-md-center">
+                                <tr>
+                                    <th scope="col">Nome do Paciente</th>
+                                    <th scope="col">Código</th>
+                                    <th scope="col">Data de criação</th>
+                                    <th scope="col">Situação</th>
+                                    <th scope="col">Funcionário</th>
+                                    <th scope="col">Ações</th>
+                                </tr>
+                            </thead>
+                            <tbody class="text-md-center">
+                                <%for(int i = 0; i < list.size(); i++){
+                                    out.println("<tr>");
+                                    out.println("<td>"+list.get(i).getNomepaciente()+"</td>");
+                                    out.println("<td>"+list.get(i).getCodigo()+"</td>");
+                                    out.println("<td>"+list.get(i).getData()+"</td>");    
+                                    out.println("<td>"+list.get(i).getUsuario().getNome()+"</td>");  
+                                    out.println("<td>"+list.get(i).getSituacao()+"</td>");  
+                                    out.println("<td>");
+                                    out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=excluir&id="+list.get(i).getId()+"'>");
+                                    out.println("<img src='img/excluir.png' title='excluir atendimento' width='20px' height='20px'>");
+                                    out.println("</a>");
+                                    out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=editar&id="+list.get(i).getId()+"'>");
+                                    out.println("<img src='img/editar.png' title='editar atendimento' width='20px' height='20px'>");
+                                    out.println("</a>");
+                                    out.println("<a class='btn btn-outline-dark' href='atendimentosController.jsp?funcao=detalhar&id="+list.get(i).getId()+"'>");
+                                    out.println("<img src='img/informacao.png' title='detalhar atendimento' width='20px' height='20px'>");
+                                    out.println("</a>");
+                                    out.println("</td>");
+                                    out.println("</tr>");
+                                }
+                }%>
                     </tbody>
                 </table>
             </div>
